+++ puffin-imgui/CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 * Add slider for controlling number of frames recorded.
 * Show total frames recorded and their total size.
+* Lower the default number of recorded frames to 600.
+* Add checkbox to toggle the profiling scopes.
 
 
 ## 0.13.4 - 2021-11-05